# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoLoop",
        "description": "Application de dons et d'échanges d'objets durables",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/images/{annonce}": {
            "post": {
                "tags": [
                    "Image"
                ],
                "summary": "Ajout d'une image sur une annonce",
                "operationId": "e791b6fac1c2fd1b8fccff1e44809d80",
                "parameters": [
                    {
                        "name": "annonce",
                        "in": "path",
                        "description": "ID de l'annonce",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Fichier de photo",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Image ajoutée avec succées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/images/{image}": {
            "delete": {
                "tags": [
                    "Image"
                ],
                "summary": "Supprimer une image",
                "operationId": "6046b06e3e8e84c999d4cea5ca838d3b",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "ID de l'image à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/localites": {
            "get": {
                "tags": [
                    "Localite"
                ],
                "summary": "liste de toutes les localités",
                "operationId": "d37e24bcb9ce1135d4f9c3a76a0dbdc7",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Localite"
                ],
                "summary": "Ajout d'une localité",
                "operationId": "9970da7addbaa7dc53e9911297d7ec41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "nom"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Localité créé avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/nombresLocalites": {
            "get": {
                "tags": [
                    "Localite"
                ],
                "summary": "Nombres de localites",
                "operationId": "0b938565ab69bfb5f397527f5d73b94d",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/localites/{localite}": {
            "put": {
                "tags": [
                    "Localite"
                ],
                "summary": "Modifier le nom d'une localité",
                "operationId": "eb76eda7fa9c591a297390fad38ee0f2",
                "parameters": [
                    {
                        "name": "localite",
                        "in": "path",
                        "description": "ID de le localite à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    " nom": {
                                        "type": "string",
                                        "example": "Nouveau nom"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Rôle non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Localite"
                ],
                "summary": "Supprimer une localité",
                "operationId": "35788788bc29fd9cdfaf851f643190ef",
                "parameters": [
                    {
                        "name": "localite",
                        "in": "path",
                        "description": "ID de le localité",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}